Added HipChat preview of graph - Still WIP but functionally works as desired.  If an error occurs at say 16:33 UTC time, the standard hipchat error message appears, but underneath now also has an absolute time graph of the last 1hour so 15:33 to 16:33. previously i had implemented this using just graphites url render api from=-1hours relative time, but this wasn't very useful if you wanted to look at the room at a later date as the -1hour view really did just that, so if the error occurred at 16:33 but you go to the room in the evening around 20:00, you would see a graph between 19:00 and 20:00 - not useful at all, so to change this, i had to take the current time of the alert and convert it to graphites url render api absolute time format, then append it to the rest of the string.
Like i said, its work in progress, i would consider moving the preview code out of hipchat service and into a util class, as like i mentioned in issue #91 i would like to see a preview option in the subscription level, then this could be used by more than other notification services if selected when creating the subscription.

The unit test for the hipchat service needs updating, this is just to show what is possible.